Adding javadocs to the example-service-book-rest.
