+++ service/example-service-book-rest/src/main/java/io/americanexpress/service/book/rest/BookApplication.java
@@ -19,6 +19,9 @@ import io.swagger.v3.oas.annotations.info.Info;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+/**
+ * BookApplication starts the Spring Boot Application for the book rest sample.
+ */
 @OpenAPIDefinition(info = @Info(
         title = "Book API",
         version = "v1.0.0",
@@ -27,11 +30,11 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class BookApplication {
 
     /**
-     * Run the SpringBoot application.
+     * Main method to run the Spring Boot Book Application.
      *
-     * @param args
+     * @param args the args
      */
-    public static void main(String args[]) {
+    public static void main(String[] args) {
         SpringApplication.run(BookApplication.class, args);
     }
