+++ service/example-service-book-rest/src/main/java/io/americanexpress/service/book/rest/config/BookConfig.java
@@ -19,13 +19,18 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.PropertySource;
 
+/**
+ * BookConfig is the configuration class for the Book Application.
+ */
 @Configuration
 @PropertySource("classpath:/service-book-application.properties")
 @ComponentScan(basePackages = "io.americanexpress.service.book.rest")
 @Import({ServiceRestConfig.class})
 public class BookConfig {
 
+    /**
+     * The constant BOOK_ENDPOINT.
+     */
     public static final String BOOK_ENDPOINT = "/v1/books";
-
 
 }
