+++ service/example-service-book-rest/src/main/java/io/americanexpress/service/book/rest/controller/ReadMonoBookController.java
@@ -8,6 +8,9 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+/**
+ * ReadMonoBookController retrieves a Book Response given its unique identifier.
+ */
 @Tag(name = "Get Book Service", description = "Provides book information based on identifier")
 @RestController
 @RequestMapping(BookConfig.BOOK_ENDPOINT)
