+++ service/example-service-book-rest/src/main/java/io/americanexpress/service/book/rest/service/ReadMonoBookService.java
@@ -17,6 +17,9 @@ import io.americanexpress.service.book.rest.model.ReadBookResponse;
 import io.americanexpress.synapse.service.rest.service.BaseGetMonoService;
 import org.springframework.stereotype.Service;
 
+/**
+ * ReadMonoBookService retrieves a Book Response given its unique identifier.
+ */
 @Service
 public class ReadMonoBookService extends BaseGetMonoService<ReadBookResponse> {
