+++ service/example-service-book-rest/src/main/java/io/americanexpress/service/book/rest/service/ReadPolyBookService.java
@@ -13,8 +13,8 @@
  */
 package io.americanexpress.service.book.rest.service;
 
-import io.americanexpress.service.book.rest.model.ReadBookRequest;
 import io.americanexpress.service.book.rest.model.ReadBookResponse;
+import io.americanexpress.service.book.rest.model.ReadPolyBookRequest;
 import io.americanexpress.synapse.service.rest.service.BaseReadPolyService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -23,11 +23,14 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * ReadPolyBookService retrieves a collection of books and the results can be filtered by some given parameters.
+ */
 @Service
-public class ReadPolyBookService extends BaseReadPolyService<ReadBookRequest, ReadBookResponse> {
+public class ReadPolyBookService extends BaseReadPolyService<ReadPolyBookRequest, ReadBookResponse> {
 
     @Override
-    protected Page<ReadBookResponse> executeRead(ReadBookRequest readBookRequest) {
+    protected Page<ReadBookResponse> executeRead(ReadPolyBookRequest readBookRequest) {
         List<ReadBookResponse> readBookResponses = new ArrayList<>();
         ReadBookResponse readBookResponse = new ReadBookResponse("Synapse", "Gabriel");
         readBookResponses.add(readBookResponse);
