+++ service/example-service-book-rest/src/test/resources/postman/SYNAPSE-EXAMPLE.postman_collection.json
@@ -33,11 +33,6 @@
 						"key": "Content-Type",
 						"value": "application/json",
 						"type": "default"
-					},
-					{
-						"key": "Accept",
-						"value": "application/json",
-						"type": "default"
 					}
 				],
 				"body": {
